Add files via upload
This is the output, the first ID of the sections combined is missed up for some reason. I checked the algorithm many times but I don't see the problem because it works for the other 7 IDs in the section resulting from the (+). Also, I overloaded the operation = as you can see in the code.